refactor: move MainApplication to infrastructure module
