+++ build.gradle.kts
@@ -12,10 +12,10 @@ java {
 		languageVersion.set(JavaLanguageVersion.of(17))
 	}
 }
-
-tasks.jar {
-	enabled = false
+tasks.getByName<Jar>("jar") {
+	enabled = true
 }
+
 
 repositories {
 	mavenCentral()
@@ -49,8 +49,6 @@ dependencies {
 	implementation("com.fasterxml.jackson.datatype:jackson-datatype-jsr310:2.12.3")
 	implementation("org.springframework.boot:spring-boot-starter-aop")
 	implementation("ca.pjer:logback-awslogs-appender:1.6.0")
-	implementation("org.flywaydb:flyway-core")
-	implementation("org.flywaydb:flyway-mysql")
 
 	compileOnly("org.projectlombok:lombok")
 	runtimeOnly("com.mysql:mysql-connector-j")
@@ -73,14 +71,4 @@ kotlin {
 	compilerOptions {
 		freeCompilerArgs.addAll("-Xjsr305=strict")
 	}
-}
-
-allOpen {
-	annotation("jakarta.persistence.Entity")
-	annotation("jakarta.persistence.MappedSuperclass")
-	annotation("jakarta.persistence.Embeddable")
-}
-
-tasks.withType<Test> {
-	useJUnitPlatform()
 }
