+++ infrastructure/src/main/kotlin/com/dobby/DobbyBackendApplication.kt
@@ -1,4 +1,4 @@
-package com.dobby.backend
+package com.dobby
 
 import com.dobby.usecase.UseCase
 import org.springframework.boot.autoconfigure.SpringBootApplication
@@ -10,8 +10,6 @@ import org.springframework.context.annotation.EnableAspectJAutoProxy
 import org.springframework.context.annotation.FilterType
 
 @ComponentScan(
-	// TODO: 모든 레이어에 대해 모듈 전환 후, basePackages 삭제
-	basePackages = ["com.dobby"],
 	includeFilters = [ComponentScan.Filter(
 		type = FilterType.ASSIGNABLE_TYPE,
 		classes = [UseCase::class]
