+++ infrastructure/src/main/resources/db/migration/V1__init.sql
