+++ infrastructure/src/main/resources/template-application-local.yml
